// src/api/models/settings-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { WorkSettings } from '../models/work-settings';
export interface SettingsResponse {

  /**
   * Language preference (de_DE or en_US)
   */
  language?: string;

  /**
   * Theme preference (light, dark, or system)
   */
  theme?: string;

  /**
   * Opaque work schedule settings stored as JSON.
   */
  workSettings?: WorkSettings;
}